Fixed the new Distiller operation, undefined iterator behavior caused lots of problems. Discovered that the procedural interface will need to be reintegrated with the scripting modules since it had, at some point, been separated from them, and now it appears foolish to have so much code duplication.

+++ Loris/src/lorisNonObj_pi.C
@@ -585,7 +585,7 @@ void synthesize( const PartialList * partials,
 		//	resize the Samplevector if necessary (pad the length
 		//	to accomodate the fade-out at the end of the latest
 		//	Partial):
-		const long nsamps = srate * ( maxtime + Partial::FadeTime() );	
+		const long nsamps = long( srate * ( maxtime + Partial::FadeTime() ) );	
 		if ( samples->size() < nsamps )
 			samples->resize( nsamps, 0. );
